Move all auth files to a new auth directory
Addresses #34
