+++ client/src/app/doorBoard/add-doorBoard.component.ts
@@ -4,7 +4,7 @@ import { MatSnackBar } from '@angular/material/snack-bar';
 import { Router } from '@angular/router';
 import { DoorBoard } from './doorBoard';
 import { DoorBoardService } from './doorBoard.service';
-import { AuthService } from '../auth.service';
+import { AuthService } from '../auth/auth.service';
 
 @Component({
   selector: 'app-add-doorBoard',
