+++ client/tsconfig.spec.json
@@ -1,6 +1,7 @@
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {
+    "experimentalDecorators": true,
     "outDir": "./out-tsc/spec",
     "types": [
       "jasmine",
